moved frame logic to MainExplore
